[add-新增功能] 功能描述:添加qps推送 影响范围: StoryID :

+++ dfas-monitor-common/src/main/java/com/win/dfas/monitor/common/vo/QpsVO.java
@@ -14,6 +14,6 @@ public class QpsVO {
 
     private List<String> xAxisData = new ArrayList<>();
 
-    private List<Integer> yAxisData = new ArrayList<>();
+    private List<Double> yAxisData = new ArrayList<>();
 
 }
